velocity calculation for hyperbolic orbits

+++ src/game_engine/physics/PhysicsEngine.h
@@ -47,7 +47,7 @@ class PhysicsEngine {
     CelestialBody * findCelestialBody(int id);
     double eccentricAnomaly(OrbitalParameters& orbitalParameters, uint64_t time, double parent_mass) const;
     double meanAnomaly(OrbitalParameters& orbitalParameters, uint64_t time, double parent_mass) const;
-    double trueAnomaly(OrbitalParameters& orbitalParameters, uint64_t time, int parent_mass) const;
+    double trueAnomaly(OrbitalParameters& orbitalParameters, uint64_t time, double parent_mass) const;
     static double averageAngularVelocity(OrbitalParameters& orbitalParameters, double parent_body_mass);
     static double period(OrbitalParameters& orbitalParameters);
     static double periapsis(OrbitalParameters& orbitalParameters) ;
